# backend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Salsyx API",
    "description": "Search and preservation platform for public GitHub repositories. Nothing open-source should disappear forever.",
    "version": "0.1.0",
    "license": { "name": "MIT" }
  },
  "servers": [
    { "url": "/api/v1", "description": "Default base path" }
  ],
...
  "paths": {
...
    "/archive/{id}": {
      "get": {
        "tags": ["archive"],
        "summary": "Get archive metadata",
        "operationId": "getArchive",
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "string", "format": "uuid" } }
        ],
        "responses": {
          "200": {
            "description": "Archive metadata",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ArchiveResponse" }
              }
            }
          }
        }
      }
    },
    "/download/{id}": {
      "get": {
        "tags": ["archive"],
        "summary": "Download archived blob",
        "description": "Streams the archived snapshot with integrity verification.",
        "operationId": "downloadArchive",
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "string", "format": "uuid" } }
        ],
        "responses": {
          "200": {
            "description": "Binary archive stream"
          }
        }
      }
    },
...
    "/archive": {
      "post": {
        "tags": ["archive"],
        "summary": "Enqueue an archive job",
        "operationId": "createArchive",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/CreateArchiveRequest" }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Archive job queued",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/CreateArchiveResponse" }
              }
            }
          }
        }
      }
    },
...
    "/repo/{owner}/{repo}/archives": {
      "get": {
        "tags": ["repository"],
        "summary": "Archive history for a repository",
        "operationId": "repoArchiveHistory",
        "parameters": [
          { "name": "owner", "in": "path", "required": true, "schema": { "type": "string" } },
          { "name": "repo", "in": "path", "required": true, "schema": { "type": "string" } }
        ],
        "responses": {
          "200": {
            "description": "Archives, newest first",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/HistoryResponse" }
              }
            }
          }
        }
      }
    },
    "/owner/{login}": {
      "get": {
        "tags": ["owner"],
        "summary": "Owner profile",
        "description": "Live GitHub profile (followers/following/public repos/pinned) plus preserved-repo counts.",
        "operationId": "getOwner",
        "parameters": [
          { "name": "login", "in": "path", "required": true, "schema": { "type": "string" } }
        ],
        "responses": {
          "200": {
            "description": "Owner profile",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/OwnerResponse" }
              }
            }
          },
          "404": {
            "description": "Owner not found on GitHub",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorBody" }
              }
            }
          }
        }
      }
    },
    "/archive/{id}/tree": {
      "get": {
        "tags": ["archive"],
        "summary": "Browse the preserved file tree",
        "operationId": "archiveTree",
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "string", "format": "uuid" } }
        ],
        "responses": {
          "200": {
            "description": "Recursive file listing",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/TreeResponse" }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ArchiveResponse": {
        "type": "object",
        "properties": {
          "archive": { "$ref": "#/components/schemas/Archive" },
          "download_url": { "type": "string" },
          "storage_provider": { "type": "string" }
        }
      },
      "Archive": {
        "type": "object",
        "properties": {
          "id": { "type": "string", "format": "uuid" },
          "repository_id": { "type": "string", "format": "uuid" },
          "commit_ref": { "type": "string" },
          "commit_count": { "type": "integer" },
          "checksum": { "type": "string" },
          "size_bytes": { "type": "integer" },
          "storage": {
            "type": "object",
            "properties": {
              "provider": { "type": "string" },
              "key": { "type": "string" }
            }
          },
          "compression": { "type": "string", "enum": ["zip", "git_bundle", "tar", "custom"] },
          "status": { "type": "string", "enum": ["pending", "fetching", "processing", "archived", "verification_failed", "failed"] },
          "deleted_at": { "type": "string", "format": "date-time" },
          "archived_at": { "type": "string", "format": "date-time" }
        }
      },
...
      "CreateArchiveRequest": {
        "type": "object",
        "required": ["full_name"],
        "properties": {
          "full_name": { "type": "string", "description": "owner/repo" }
        }
      },
      "CreateArchiveResponse": {
        "type": "object",
        "properties": {
          "archive_id": { "type": "string", "format": "uuid" },
          "status": { "type": "string" },
          "message": { "type": "string" }
        }
      },
...
      "OwnerTopRepo": {
        "type": "object",
        "properties": {
          "name": { "type": "string" },
          "full_name": { "type": "string" },
          "description": { "type": "string" },
          "language": { "type": "string" },
          "stars_count": { "type": "integer" },
          "html_url": { "type": "string" }
        }
      },
      "PinnedRepo": {
        "type": "object",
        "properties": {
          "name": { "type": "string" },
          "full_name": { "type": "string" },
          "description": { "type": "string" },
          "primary_language": { "type": "string" },
          "stargazer_count": { "type": "integer" },
          "html_url": { "type": "string" }
        }
      },
      "OwnerResponse": {
        "type": "object",
        "properties": {
          "login": { "type": "string" },
          "name": { "type": "string" },
          "bio": { "type": "string" },
          "avatar_url": { "type": "string" },
          "html_url": { "type": "string" },
          "followers": { "type": "integer" },
          "following": { "type": "integer" },
          "public_repos": { "type": "integer" },
          "preserved_count": { "type": "integer" },
          "top_repos": { "type": "array", "items": { "$ref": "#/components/schemas/OwnerTopRepo" } },
          "pinned_repos": { "type": "array", "items": { "$ref": "#/components/schemas/PinnedRepo" } }
        }
      },
      "TreeEntry": {
        "type": "object",
        "properties": {
          "path": { "type": "string" },
          "type": { "type": "string", "enum": ["blob", "tree", "commit"] },
          "size": { "type": "integer" }
        }
      },
      "TreeResponse": {
        "type": "object",
        "properties": {
          "archive_id": { "type": "string", "format": "uuid" },
          "commit_ref": { "type": "string" },
          "entries": { "type": "array", "items": { "$ref": "#/components/schemas/TreeEntry" } },
          "regenerated": { "type": "boolean" }
        }
      },
      "HistoryArchive": {
        "type": "object",
        "properties": {
          "id": { "type": "string", "format": "uuid" },
          "commit_ref": { "type": "string" },
          "commit_count": { "type": "integer" },
          "checksum": { "type": "string" },
          "size_bytes": { "type": "integer" },
          "status": { "type": "string" },
          "archived_at": { "type": "string", "format": "date-time" }
        }
      },
      "HistoryResponse": {
        "type": "object",
        "properties": {
          "owner": { "type": "string" },
          "repo": { "type": "string" },
          "archives": { "type": "array", "items": { "$ref": "#/components/schemas/HistoryArchive" } }
        }
      },
...
      "ErrorBody": {
        "type": "object",
        "properties": {
          "code": { "type": "string" },
          "message": { "type": "string" },
          "detail": { "type": "string" }
        }
      }
    }
  }
}